microservices/analytics-service/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
google-cloud-storage==2.14.0
prometheus-client==0.19.0
gunicorn==21.2.0

